+++ src/Serilog.Sinks.AzureEventHub/Sinks/AzureEventHub/AzureEventHubSink.cs
@@ -14,8 +14,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Text;
 using Azure.Messaging.EventHubs;
 using Azure.Messaging.EventHubs.Producer;
 using Serilog.Core;
@@ -31,18 +29,22 @@ namespace Serilog.Sinks.AzureEventHub
     {
         readonly EventHubProducerClient _eventHubClient;
         readonly ITextFormatter _formatter;
+        readonly bool _shouldIncludeProperties;
 
         /// <summary>
         /// Construct a sink that saves log events to the specified EventHubClient.
         /// </summary>
         /// <param name="eventHubClient">The EventHubClient to use in this sink.</param>
         /// <param name="formatter">Provides formatting for outputting log data</param>
+        /// <param name="shouldIncludeProperties">Whether the log event's properties are added to the EventData.</param>
         public AzureEventHubSink(
             EventHubProducerClient eventHubClient,
-            ITextFormatter formatter)
+            ITextFormatter formatter,
+            bool shouldIncludeProperties = false)
         {
             _eventHubClient = eventHubClient;
             _formatter = formatter;
+            _shouldIncludeProperties = shouldIncludeProperties;
         }
 
         /// <summary>
@@ -51,15 +53,7 @@ namespace Serilog.Sinks.AzureEventHub
         /// <param name="logEvent">The log event to write.</param>
         public void Emit(LogEvent logEvent)
         {
-            byte[] body;
-            using (var render = new StringWriter())
-            {
-                _formatter.Format(logEvent, render);
-                body = Encoding.UTF8.GetBytes(render.ToString());
-            }
-            var eventHubData = new EventData(body);
-            eventHubData.Properties.Add("Type", "SerilogEvent");
-            eventHubData.Properties.Add("Level", logEvent.Level.ToString());
+            var eventHubData = EventDataFactory.CreateEventData(logEvent, _formatter, _shouldIncludeProperties);
 
             //Unfortunately no support for async in Serilog yet
             //https://github.com/serilog/serilog/issues/134